+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 =========
 
 * :bug:`-` When ``verbose=False`` set the ``StreamHandler`` level to ``ERROR`` to allow tracebacks.
+* :feature:`11` *Breaking change.* `.Exposure.write` is now a coroutine and must be awaited if called directly. ``HDUList.writeto()`` is run in an executor.
 
 * :release:`0.2.0 <2020-07-31>`
 * :bug:`-` Fix ``CameraWarning`` when used from a ``CameraSystem`` instance.